requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Werkzeug==3.0.6
Jinja2==3.1.6
click==8.1.8
itsdangerous==2.2.0
blinker==1.8.2
pytest==8.0.0
pytest-cov==4.1.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.27
faiss-cpu>=1.7.4
openai>=1.0.0
tiktoken==0.5.2
langchain-openai>=0.0.2
sentence-transformers>=2.2.2
torch>=2.0.0
ollama>=0.1.6
huggingface-hub>=0.19.4 